api_tests/requirements.txt
pytest
requests
pytest-cov
pytest-html
jsonschema
pytest-xdist>=3.6.1
python-dotenv
